전략패턴 - 전략 패턴으로 코드 분리
