+++ youngchul.shin/ch07/example/src/main/java/study/pattern/discount/DiscountStrategy.java
@@ -1,0 +1,8 @@
+package study.pattern.discount;
+
+import study.pattern.Item;
+
+public interface DiscountStrategy {
+
+  int getDiscountPrice(Item item);
+}
