+++ youngchul.shin/ch07/example/src/main/java/study/pattern/discount/FirstGuestDiscountStrategy.java
@@ -1,0 +1,11 @@
+package study.pattern.discount;
+
+import study.pattern.Item;
+
+public class FirstGuestDiscountStrategy implements DiscountStrategy {
+
+  @Override
+  public int getDiscountPrice(Item item) {
+    return (int)(item.getPrice() * 0.9);
+  }
+}
