+++ youngchul.shin/ch07/example/src/main/java/study/pattern/discount/NonFreshItemDiscountStrategy.java
@@ -1,0 +1,11 @@
+package study.pattern.discount;
+
+import study.pattern.Item;
+
+public class NonFreshItemDiscountStrategy implements DiscountStrategy {
+
+  @Override
+  public int getDiscountPrice(Item item) {
+    return (int)(item.getPrice() * 0.8);
+  }
+}
